+++ xbmc/cores/AudioEngine/Engines/PulseAEStream.h
@@ -37,9 +37,9 @@ public:
 
   virtual unsigned int GetSpace();
   virtual unsigned int AddData(void *data, unsigned int size);
-  virtual float GetDelay();
-  virtual float GetCacheTime ();
-  virtual float GetCacheTotal();
+  virtual double GetDelay();
+  virtual double GetCacheTime ();
+  virtual double GetCacheTotal();
 
   virtual bool IsPaused     ();
   virtual bool IsDraining   ();
